newsLetter/requirements.txt
spacy
llama-index
streamlit
reportlab
openai
nltk
GoogleNews
newspaper3k
